# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mahasiswa",
        "description": "Dokumentasi API Mahasiswa dengan Bearer Token",
        "version": "1.0.0"
    },
    "paths": {
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil semua data mahasiswa",
                "operationId": "dbf3137c7d1805b9cf58fe27b70a6e89",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah data mahasiswa",
                "operationId": "66fe54445ec2d86b3347d43216a5045f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "nama_mahasiswa",
                                    "fakultas",
                                    "jurusan"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "nama_mahasiswa": {
                                        "type": "string",
                                        "example": "Rizal"
                                    },
                                    "fakultas": {
                                        "type": "string",
                                        "example": "Teknik"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Informatika"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menambahkan data"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil data mahasiswa berdasarkan ID",
                "operationId": "74de419ee2bf151bc82b9044e182df7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Perbarui data mahasiswa",
                "operationId": "ae894e88f2aa13240bc5ce5d3df51774",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_mahasiswa": {
                                        "type": "string",
                                        "example": "Rizal Baru"
                                    },
                                    "fakultas": {
                                        "type": "string",
                                        "example": "Ekonomi"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Manajemen"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus data mahasiswa",
                "operationId": "416b66a5d2acf8515b2277b8c4d2304a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus data"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Mahasiswa",
            "description": "Mahasiswa"
        }
    ]
}